moved the gridded data tests to a new file, full smrf tests now clean up after themselves, hrrr gold albedo defaults were different
